refactor(recruitment): extract public-shortcode renderer (S7 of #141)
Pure code movement — no behavior changes. RecruitmentPublicShortcode
drops from 1136 to 476 LOC after extracting all the badge/label/
format/layout helpers into a focused renderer.

  - RecruitmentPublicShortcodeRenderer (692 LOC) — render_section,
    render_row, render_filters_bar, render_subscription_filter_inputs,
    render_name_search_input, render_adjutancy_filter_inputs,
    render_pagination, parse_columns_config, wrap_with_banner,
    decrypt_field, status_label, render_*_badge methods,
    preview_status_label, format_date_br, format_time_hm, msg.

Visibility: 5 methods called from the facade (render_uncached, render)
are public — render_section, render_filters_bar, parse_columns_config,
wrap_with_banner, msg. The remaining 14 helpers stay private since
they're only called from sibling renderer methods.

Facade keeps the public API + cache/rate-limit infrastructure:
register, render, render_uncached, wrap_output, enqueue_public_css,
cache_key, invalidate_public_cache, check_rate_limit, client_ip, plus
all SHORTCODE_TAG / CACHE_PREFIX / RATE_PREFIX / CACHE_VERSION_OPTION
/ CACHE_DIRTY_ACTION constants. Inside render_uncached() and render(),
the 5 calls to moved helpers became
RecruitmentPublicShortcodeRenderer::method(...). Unused use imports
for BadgeHtml, DocumentFormatter, and Encryption migrated to the
renderer where they're now used.

Verification:
  - php -l: clean on both files.
  - grep self:: in facade: only references own methods/constants.
  - vendor/bin/phpunit --filter RecruitmentPublicShortcode: 10/10 OK.
  - vendor/bin/phpunit: 3870/3870 OK.
  - vendor/bin/phpstan analyse includes/recruitment/: no errors.
